@@ -3,30 +3,32 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/react": "^12.1.3",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.4.1",
-    "@types/react": "^18.3.18",
-    "@types/react-dom": "^18.3.5",
-    "axios": "^1.5.1",
-    "formik": "^2.4.5",
-    "i18next": "^21.6.14",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-i18next": "^11.16.2",
+    "@tanstack/react-query": "^5.67.1",
+    "@testing-library/react": "^16.2.0",
+    "@testing-library/user-event": "^14.6.1",
+    "@types/jest": "^29.5.14",
+    "@types/react": "^19.0.10",
+    "@types/react-dom": "^19.0.4",
+    "axios": "^1.8.1",
+    "eslint-plugin-react-refresh": "^0.4.19",
+    "formik": "^2.4.6",
+    "i18next": "^24.2.2",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "react-i18next": "^15.4.1",
     "react-icons": "^5.5.0",
-    "react-router-dom": "^6.18.0",
-    "styled-components": "^5.3.3",
-    "typescript": "^4.6.2",
-    "web-vitals": "^2.1.4",
-    "yup": "^1.3.2"
+    "react-router-dom": "^7.2.0",
+    "styled-components": "^6.1.15",
+    "typescript": "^5.8.2",
+    "web-vitals": "^4.2.4",
+    "yup": "^1.6.1"
   },
   "scripts": {
     "dev": "vite",
     "build": "vite build",
     "preview": "vite preview",
-    "lint": "eslint src --ext .js,.jsx,.ts,.tsx -c ./.eslintrc.json",
-    "lint:fix": "eslint src --ext .js,.jsx,.ts,.tsx -c ./.eslintrc.json --fix"
+    "lint": "eslint src -c ./eslint.config.mjs",
+    "lint:fix": "eslint src -c ./eslint.config.mjs --fix"
   },
   "browserslist": {
     "production": [
@@ -41,19 +43,21 @@
     ]
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.16.2",
-    "@types/node": "^20.17.23",
-    "@types/styled-components": "^5.1.24",
-    "@typescript-eslint/eslint-plugin": "^5.13.0",
-    "@typescript-eslint/parser": "^5.13.0",
+    "@eslint/js": "^9.21.0",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@types/node": "^22.13.9",
+    "@types/styled-components": "^5.1.34",
+    "@typescript-eslint/eslint-plugin": "^8.26.0",
+    "@typescript-eslint/parser": "^8.26.0",
     "@vitejs/plugin-react": "^4.3.4",
-    "eslint": "^8.10.0",
-    "eslint-config-airbnb": "^19.0.4",
-    "eslint-import-resolver-typescript": "^2.5.0",
-    "eslint-plugin-import": "^2.25.4",
-    "eslint-plugin-jsx-a11y": "^6.5.1",
-    "eslint-plugin-react": "^7.29.3",
-    "eslint-plugin-react-hooks": "^4.3.0",
+    "eslint": "^8.57.1",
+    "eslint-import-resolver-typescript": "^3.8.3",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-jsx-a11y": "^6.10.2",
+    "eslint-plugin-react": "^7.37.4",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "globals": "^16.0.0",
+    "typescript-eslint": "^8.26.0",
     "vite": "^6.2.0"
   }
 }
